Split pybind deprecation test
Separate the autocompletion portion of deprecation_test.py (from
bindings/pydrake/common), as it causes strange things to happen to the
module's reference count under Jammy (Python 3.10), resulting in that
portion of the test failing.
