+++ bindings/pydrake/common/BUILD.bazel
@@ -559,11 +559,22 @@ drake_py_library(
 )
 
 # Note: This target tests the low-level deprecation API.
-# Please see `deprecation_utility_test` for a unittest on higher-level
-# deprecation API.
+# See `deprecation_utility_test` for a unittest on higher-level deprecation
+# API.
 drake_py_unittest(
     name = "deprecation_test",
     tags = ["no_kcov"],  # kcov messes with module ref counts.
+    deps = [
+        # N.B. We include this here due to awkward setup for `pydrake.common`.
+        ":module_py",
+        ":deprecation_example",
+    ],
+)
+
+# Note: This target tests autocompletion for the low-level deprecation API.
+# See source for an explanation why this is separate from `deprecation_test`.
+drake_py_unittest(
+    name = "deprecation_autocomplete_test",
     deps = [
         # N.B. We include this here due to awkward setup for `pydrake.common`.
         ":module_py",
