+++ bindings/pydrake/common/test/deprecation_test.py
@@ -1,10 +1,11 @@
 """
 Tests low-level deprecation API.
 
-Please see `deprecation_utility_test.py` for a unittest on higher-level API.
+See `deprecation_utility_test.py` for a unittest on higher-level API. See
+`deprecation_autocomplete_test.py` for an autocompletion test of this API, as
+well as an explanation why that test is separate.
 """
 
-import rlcompleter
 import sys
 from types import ModuleType
 import unittest
@@ -12,22 +13,6 @@ import warnings
 
 from pydrake.common.deprecation import (
     DrakeDeprecationWarning, _forward_callables_as_deprecated)
-
-
-def get_completion_suffixes(namespace, prefix, max_count=1000):
-    # Gets all completions for a given namespace and prefix, stripping the
-    # prefix from the results.
-    completer = rlcompleter.Completer(namespace)
-    suffixes = []
-    for i in range(max_count):
-        candidate = completer.complete(prefix, i)
-        if candidate is None:
-            break
-        assert candidate.startswith(prefix), (prefix, candidate)
-        suffixes.append(candidate[len(prefix):])
-    else:
-        raise RuntimeError("Exceeded max count!")
-    return suffixes
 
 
 class TestDeprecation(unittest.TestCase):
@@ -83,61 +68,6 @@ class TestDeprecation(unittest.TestCase):
         temp = {}
         exec("from deprecation_example import *", temp, temp)
         self.assertIsInstance(temp["sub_module"], str)
-
-    def test_module_autocomplete(self):
-        # Ensure that we can autocomplete with our example module.
-        # Without `__dir__` being implemented, it'll only return `install` as a
-        # non-private autocomplete candidate.
-        import deprecation_example
-        suffixes = get_completion_suffixes(
-            locals(), prefix="deprecation_example.")
-        suffixes_expected = [
-            # Injection from `Completer.attr_matches`, via `get_class_members`.
-            "__class__(",
-            "__delattr__(",
-            "__dict__",
-            "__dir__(",
-            "__doc__",
-            "__format__(",
-            "__getattr__(",
-            "__getattribute__(",
-            "__hash__(",
-            "__init__(",
-            "__module__",
-            "__new__(",
-            "__reduce__(",
-            "__reduce_ex__(",
-            "__repr__(",
-            "__setattr__(",
-            "__sizeof__(",
-            "__str__(",
-            "__subclasshook__(",
-            "__weakref__",
-            "_install(",
-            # Intended completions via `__all__`.
-            "sub_module",
-            "value",
-        ]
-        suffixes_expected += [
-            "__ge__(",
-            "__eq__(",
-            "__le__(",
-            "__lt__(",
-            "__gt__(",
-            "__ne__(",
-        ]
-        if hasattr(deprecation_example, "__init_subclass__"):
-            suffixes_expected.append("__init_subclass__(")
-        # For Bionic Python3, the behavior of autocompletion seems to
-        # constrain behavior depending on underscore prefixes.
-        if "__init__(" not in suffixes:
-            under = get_completion_suffixes(
-                locals(), prefix="deprecation_example._")
-            suffixes += ["_" + s for s in under]
-            dunder = get_completion_suffixes(
-                locals(), prefix="deprecation_example.__")
-            suffixes += ["__" + s for s in dunder]
-        self.assertSetEqual(set(suffixes), set(suffixes_expected))
 
     def _check_warning(self, item, message_expected, check_full=True):
         self.assertEqual(item.category, DrakeDeprecationWarning)
